requirements.txt
Flask==2.1.1
google-generativeai==0.2.0
langchain-google-genai==0.1.0
langchain-community-vectorstores==0.1.0
faiss-cpu==1.7.3
flask-cors==3.0.10
opencv-python-headless==4.5.5.64
numpy==1.21.2
fer==21.0.5